+++ packages/frontend/.storybook/main.js
@@ -3,6 +3,7 @@ const { TsconfigPathsPlugin } = require('tsconfig-paths-webpack-plugin');
 module.exports = {
   stories: ['../**/*.stories.mdx', '../**/*.stories.@(js|jsx|ts|tsx)'],
   addons: [
+    '@storybook/preset-scss',
     '@storybook/addon-links',
     '@storybook/addon-essentials',
     '@storybook/addon-a11y',